Share planned pipeline body assembly
