+++ jetro-core/src/exec/pipeline/logical_lower.rs
@@ -8,9 +8,7 @@
 use std::sync::Arc;
 
 use crate::builtins::{BuiltinMethod, BuiltinViewStage};
-use crate::exec::pipeline::{
-    plan_with_exprs, Pipeline, PipelineBody, ReducerSpec, Sink, Source, Stage,
-};
+use crate::exec::pipeline::{Pipeline, PipelineBody, ReducerSpec, Sink, Source, Stage};
 use crate::ir::logical::LogicalPlan;
 use crate::parse::ast::Expr;
 
@@ -200,22 +198,9 @@ fn collect_numeric_sink(
 // Body assembly
 // ---------------------------------------------------------------------------
 
-/// Classifies stages into `BodyKernel`s, runs `plan_with_exprs` for filter reordering/fusion,
-/// and fills in the kernel vectors required by `PipelineBody`.
+/// Runs the shared pipeline planner and fills in kernel vectors required by `PipelineBody`.
 fn build_body(stages: Vec<Stage>, stage_exprs: Vec<Option<Arc<Expr>>>, sink: Sink) -> PipelineBody {
-    let kernels = Stage::body_kernels(&stages);
-    let plan_result = plan_with_exprs(stages, stage_exprs, &kernels, sink);
-
-    let stage_kernels = Stage::body_kernels(&plan_result.stages);
-    let sink_kernels = plan_result.sink.body_kernels();
-
-    PipelineBody {
-        stages: plan_result.stages,
-        stage_exprs: plan_result.stage_exprs,
-        sink: plan_result.sink,
-        stage_kernels,
-        sink_kernels,
-    }
+    PipelineBody::planned(stages, stage_exprs, sink)
 }
 
 // ---------------------------------------------------------------------------
